pyproject.toml
[project]
name = "adk-sample-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "google-adk>=1.18.0",
  "pip>=25.3"
]
